requirements.txt
annotated-types
anyio
bcrypt==3.2.0
certifi
cffi
charset-normalizer
click
colorama
cryptography
dnspython
ecdsa
email-validator
fastapi
filelock
fsspec
greenlet
h11
huggingface-hub
idna
Jinja2
joblib
MarkupSafe
mpmath
networkx
nltk
numpy
packaging
passlib
psycopg2-binary
pyasn1
pycparser
pydantic
pydantic_core
python-dotenv
python-jose
PyJWT
PyYAML
regex
requests
rsa
safetensors
setuptools
six
sniffio
SQLAlchemy
starlette
sympy
textblob
tokenizers
torch
tqdm
transformers
typing-inspection
typing_extensions
urllib3
uvicorn
SpeechRecognition
soundfile
standard-aifc
python-multipart
motor
vaderSentiment
google-generativeai
